API for Selective table
Added a temporary table for selective (sorting is disabled for now)

Added support for the selective table in the topTable.js, only temporary as it should be seperated to a different file

Added support for display of all of the subject of a threads ids array. Currently temporarly using only the top sender emails as an example.