Added a trash icon for the checkboxes.
Added event handlers for all of the rows so that if the trash icon will be displayed only if 1 or more checkboxs are marked. if all of the checkboxes are unmarked, the icon will be removed. (need to verify that all of the event handlers are removed after table reload).

Added event handler to the trash icon, so that it will sum up all of the threads that are marked (still need to handle the concatination where there is a missing comma).